+++ Sources/Features/Application List/ApplicationListItem.swift
@@ -1,78 +1,5 @@
 import Cocoa
 
-class OpaqueView: NSView {
-  override var isOpaque: Bool { return true }
-}
-
-class TextField: NSTextField {
-  override var isOpaque: Bool { return true }
-}
-
-class Spacer: NSView {
-  override var isOpaque: Bool { return true }
-
-  convenience init(size: CGFloat) {
-    self.init(frame: .init(origin: .zero, size: .init(width: size, height: size)))
-  }
-
-  override func viewWillMove(toSuperview newSuperview: NSView?) {
-    super.viewWillMove(toSuperview: newSuperview)
-    setContentHuggingPriority(.required, for: .horizontal)
-    setContentHuggingPriority(.required, for: .vertical)
-  }
-}
-
-class ImageView: NSImageView {
-  override var isOpaque: Bool { return true }
-
-  override init(frame frameRect: NSRect) {
-    super.init(frame: frameRect)
-    wantsLayer = true
-  }
-
-  required init?(coder: NSCoder) {
-    fatalError("init(coder:) has not been implemented")
-  }
-}
-
-class StackView: NSStackView {
-  override var isOpaque: Bool { return true }
-
-  convenience init(_ views: [NSView]) {
-    self.init(views: views)
-  }
-
-  convenience init(_ views: NSView...) {
-    self.init(views: views)
-  }
-}
-
-class HStack: StackView {
-  override var isOpaque: Bool { return true }
-
-  override init(frame frameRect: NSRect) {
-    super.init(frame: frameRect)
-    orientation = .horizontal
-  }
-
-  required init?(coder decoder: NSCoder) {
-    fatalError("init(coder:) has not been implemented")
-  }
-}
-
-class VStack: StackView {
-  override var isOpaque: Bool { return true }
-
-  override init(frame frameRect: NSRect) {
-    super.init(frame: frameRect)
-    orientation = .vertical
-  }
-
-  required init?(coder decoder: NSCoder) {
-    fatalError("init(coder:) has not been implemented")
-  }
-}
-
 // sourcery: let application = "Application"
 class ApplicationListItem: CollectionViewItem, CollectionViewItemComponent {
   let baseView = NSView()
@@ -110,7 +37,7 @@ class ApplicationListItem: CollectionViewItem, CollectionViewItemComponent {
     syncView.layer?.drawsAsynchronously = true
     iconView.layer?.drawsAsynchronously = true
 
-    let verticalStackView = VStack(Spacer(size: 10), titleLabel, subtitleLabel, Spacer(size: 10))
+    let verticalStackView = VStack(SpacerView(size: 10), titleLabel, subtitleLabel, SpacerView(size: 10))
     verticalStackView.alignment = .leading
     verticalStackView.distribution = .equalCentering
     verticalStackView.spacing = 0
