+++ Sources/Views/ImageView.swift
@@ -1,0 +1,14 @@
+import Cocoa
+
+class ImageView: NSImageView {
+  override var isOpaque: Bool { return true }
+
+  override init(frame frameRect: NSRect) {
+    super.init(frame: frameRect)
+    wantsLayer = true
+  }
+
+  required init?(coder: NSCoder) {
+    fatalError("init(coder:) has not been implemented")
+  }
+}
