+++ Sources/Views/OpaqueView.swift
@@ -1,0 +1,5 @@
+import Cocoa
+
+class OpaqueView: NSView {
+  override var isOpaque: Bool { return true }
+}
