+++ Sources/Views/TextField.swift
@@ -1,0 +1,5 @@
+import Cocoa
+
+class TextField: NSTextField {
+  override var isOpaque: Bool { return true }
+}
